+++ pytorch_lightning/core/lightning.py
@@ -1158,6 +1158,10 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
                     optimizer.step()
                     optimizer.zero_grad()
 
+        Note:
+            If you also override the :meth:`~pytorch_lightning.core.hooks.ModelHooks.on_before_zero_grad`
+            model hook don't forget to add the call to it before ``optimizer.zero_grad()`` yourself.
+
         """
         if self.trainer.use_tpu and XLA_AVAILABLE:
             xm.optimizer_step(optimizer)
@@ -1165,6 +1169,9 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
             optimizer.step(second_order_closure)
         else:
             optimizer.step()
+
+        # model hook
+        self.on_before_zero_grad(optimizer)
 
         # clear gradients
         optimizer.zero_grad()