+++ pytorch_lightning/loggers/tensorboard.py
@@ -1,3 +1,8 @@
+"""
+TensorBoard
+-----------
+"""
+
 import csv
 import os
 from argparse import Namespace
@@ -14,26 +19,25 @@ from pytorch_lightning import _logger as log
 
 class TensorBoardLogger(LightningLoggerBase):
     r"""
-    Log to local file system in TensorBoard format
-
-    Implemented using :class:`torch.utils.tensorboard.SummaryWriter`. Logs are saved to
-    ``os.path.join(save_dir, name, version)``
+    Log to local file system in `TensorBoard <https://www.tensorflow.org/tensorboard>`_ format.
+    Implemented using :class:`~torch.utils.tensorboard.SummaryWriter`. Logs are saved to
+    ``os.path.join(save_dir, name, version)``. This is the default logger in Lightning, it comes
+    preinstalled.
 
     Example:
-        .. code-block:: python
-
-            logger = TensorBoardLogger("tb_logs", name="my_model")
-            trainer = Trainer(logger=logger)
-            trainer.train(model)
+        >>> from pytorch_lightning import Trainer
+        >>> from pytorch_lightning.loggers import TensorBoardLogger
+        >>> logger = TensorBoardLogger("tb_logs", name="my_model")
+        >>> trainer = Trainer(logger=logger)
 
     Args:
         save_dir: Save directory
-        name: Experiment name. Defaults to "default".  If it is the empty string then no per-experiment
+        name: Experiment name. Defaults to ``'default'``. If it is the empty string then no per-experiment
             subdirectory is used.
         version: Experiment version. If version is not specified the logger inspects the save
             directory for existing versions, then automatically assigns the next available version.
             If it is a string then it is used as the run-specific subdirectory name,
-            otherwise version_${version} is used.
+            otherwise ``'version_${version}'`` is used.
         \**kwargs: Other arguments are passed directly to the :class:`SummaryWriter` constructor.
 
     """
@@ -57,8 +61,8 @@ class TensorBoardLogger(LightningLoggerBase):
     def root_dir(self) -> str:
         """
         Parent directory for all tensorboard checkpoint subdirectories.
-        If the experiment name parameter is None or the empty string, no experiment subdirectory is used
-        and checkpoint will be saved in save_dir/version_dir
+        If the experiment name parameter is ``None`` or the empty string, no experiment subdirectory is used
+        and the checkpoint will be saved in "save_dir/version_dir"
         """
         if self.name is None or len(self.name) == 0:
             return self.save_dir
@@ -68,9 +72,9 @@ class TensorBoardLogger(LightningLoggerBase):
     @property
     def log_dir(self) -> str:
         """
-        The directory for this run's tensorboard checkpoint.  By default, it is named 'version_${self.version}'
-        but it can be overridden by passing a string value for the constructor's version parameter
-        instead of None or an int
+        The directory for this run's tensorboard checkpoint. By default, it is named
+        ``'version_${self.version}'`` but it can be overridden by passing a string value
+        for the constructor's version parameter instead of ``None`` or an int.
         """
         # create a pseudo standard path ala test-tube
         version = self.version if isinstance(self.version, str) else f"version_{self.version}"
@@ -80,14 +84,14 @@ class TensorBoardLogger(LightningLoggerBase):
     @property
     def experiment(self) -> SummaryWriter:
         r"""
+        Actual tensorboard object. To use TensorBoard features in your
+        :class:`~pytorch_lightning.core.lightning.LightningModule` do the following.
 
-         Actual tensorboard object. To use tensorboard features do the following.
+        Example::
 
-         Example::
+            self.logger.experiment.some_tensorboard_function()
 
-             self.logger.experiment.some_tensorboard_function()
-
-         """
+        """
         if self._experiment is not None:
             return self._experiment